adjust unit tests
